pyFANTOM/geom/CUDA/_filters.py
from ...core.CUDA._filter import (apply_filter_3D_cuda,
                                 apply_filter_2D_cuda,
                                 get_filter_weights_2D_cuda,
                                 get_filter_weights_3D_cuda,
                                 apply_filter_2D_transpose_cuda,
                                 apply_filter_3D_transpose_cuda)

from ._mesh import CuStructuredMesh2D, CuStructuredMesh3D, CuGeneralMesh
import cupy as cp
import numpy as np
from ..commons._filters import FilterKernel
from ..CPU._filters import filter_kernel_2D_general, filter_kernel_3D_general

class CuStructuredFilter3D(FilterKernel):
    """
    CUDA-accelerated 3D density filter for structured meshes.
    
    GPU version of StructuredFilter3D using CuPy. All filtering operations on GPU
    for maximum performance in 3D topology optimization. Identical API to CPU version
    but operates on GPU memory.
    
    Parameters
    ----------
    mesh : CuStructuredMesh3D
        CUDA 3D structured mesh
    r_min : float
        Filter radius in element units (e.g., r_min=1.5 includes elements within 1.5 element widths)
        
    Attributes
    ----------
    shape : tuple
        Filter matrix dimensions (n_elements, n_elements)
    weights : cupy.ndarray
        Precomputed filter weights for neighbor elements on GPU
    offsets : cupy.ndarray
        Element index offsets for neighbors within filter radius on GPU
    normalizer : cupy.ndarray
        Per-element normalization factors for adjoint operation on GPU
        
    Methods
    -------
    dot(rho)
        Apply forward filter: filtered_rho = H @ rho
    _rmatvec(sens)
        Apply adjoint filter for sensitivity backpropagation: H^T @ sens
        
    Notes
    -----
    - Filter uses cone-shaped kernel: w(d) = (r_min - d) / r_min for d < r_min
    - Automatically handles anisotropic elements (dx != dy != dz)
    - Adjoint includes proper normalization for optimization gradients
    - Memory-efficient: stores only unique weights and offsets, not full matrix
    - All arrays stored on GPU as CuPy arrays
    - 5-10x faster than CPU for large 3D problems
    - Essential for large-scale 3D optimization
    - Requires CUDA-capable GPU and CuPy
    
    Examples
    --------
    >>> from pyFANTOM.CUDA import StructuredMesh3D, StructuredFilter3D
    >>> mesh = StructuredMesh3D(nx=64, ny=64, nz=32, lx=1.0, ly=1.0, lz=0.5)
    >>> filter = StructuredFilter3D(mesh=mesh, r_min=1.5)
    >>> rho_filtered = filter.dot(rho_raw)
    """
    def __init__(self, mesh: CuStructuredMesh3D, r_min):
        super().__init__()
        self.nelx = mesh.nel[0]
        self.nely = mesh.nel[1]
        self.nelz = mesh.nel[2]
        self.r_min = r_min
        self.shape = (self.nelx * self.nely * self.nelz, self.nelx * self.nely * self.nelz)
        
        self.dtype = mesh.dtype
        
        dx, dy, dz = mesh.dx, mesh.dy, mesh.dz
        self.scales = cp.array([dx, dy, dz], dtype=self.dtype)
        self.scales = self.scales / self.scales.min()
        
        filter_rad = r_min
        n_neighbours = int(np.ceil(filter_rad))
        offset_range = cp.arange(-n_neighbours, n_neighbours + 1, dtype=cp.int32)
        
        a, b, c = cp.meshgrid(offset_range, offset_range, offset_range, indexing='ij')
        offsets = cp.vstack([a.ravel(), b.ravel(), c.ravel()]).T
        offsets_adjusted = offsets * self.scales[None]
        
        distances = cp.linalg.norm(offsets_adjusted, axis=1)
        weights = (r_min - distances) / r_min
        valid_mask = weights > 0
        offsets = offsets[valid_mask]
        weights = weights[valid_mask]
        weights /= weights.sum()
        
        self.weights = cp.array(weights, dtype=self.dtype)
        self.offsets = cp.array(offsets, dtype=cp.int32)
        
        self.normalizer = cp.zeros(self.shape[0], dtype=self.dtype)
        get_filter_weights_3D_cuda(self.nelx, self.nely, self.nelz, self.offsets, self.weights, self.normalizer)

    def _matvec(self, rho):
        v_out = cp.zeros_like(rho)
        apply_filter_3D_cuda(rho, v_out, self.nelx, self.nely, self.nelz, self.offsets, self.weights)
        return v_out

    def _rmatvec(self, rho):
        v_out = cp.zeros_like(rho)
        apply_filter_3D_transpose_cuda(rho, v_out, self.nelx, self.nely, self.nelz, self.offsets, self.weights, self.normalizer)
        return v_out


class CuStructuredFilter2D(FilterKernel):
    """
    CUDA-accelerated 2D density filter for structured meshes.
    
    GPU version of StructuredFilter2D using CuPy. Fast convolution-based filtering on GPU.
    Identical API to CPU version but operates on GPU memory for maximum performance.
    
    Parameters
    ----------
    mesh : CuStructuredMesh2D
        CUDA 2D structured mesh
    r_min : float
        Filter radius in element units
        
    Attributes
    ----------
    shape : tuple
        Filter matrix dimensions (n_elements, n_elements)
    weights : cupy.ndarray
        Precomputed filter weights on GPU
    offsets : cupy.ndarray
        Element index offsets for neighbors on GPU
    normalizer : cupy.ndarray
        Per-element normalization for adjoint on GPU
        
    Methods
    -------
    dot(rho)
        Forward filter application
    _rmatvec(sens)
        Adjoint filter for sensitivity backpropagation
        
    Notes
    -----
    - Uses cone filter: w(d) = (r_min - d) / r_min
    - Handles non-square elements (dx != dy) automatically
    - Typical r_min values: 1.5-3.0 element units
    - Larger r_min produces smoother, coarser designs
    - All arrays stored on GPU as CuPy arrays
    - GPU-accelerated filtering for maximum performance
    - Requires CUDA-capable GPU and CuPy
    
    Examples
    --------
    >>> from pyFANTOM.CUDA import StructuredMesh2D, StructuredFilter2D
    >>> mesh = StructuredMesh2D(nx=128, ny=64, lx=2.0, ly=1.0)
    >>> filter = StructuredFilter2D(mesh=mesh, r_min=2.0)
    >>> rho_smooth = filter.dot(rho)
    """
    def __init__(self, mesh: CuStructuredMesh2D, r_min):
        super().__init__()
        self.nelx = int(mesh.nelx)
        self.nely = int(mesh.nely)
        self.r_min = r_min
        self.shape = (self.nelx * self.nely, self.nelx * self.nely)
        self.dtype = mesh.dtype
        
        dx, dy = mesh.dx, mesh.dy
        self.scales = cp.array([dx, dy], dtype=self.dtype)
        self.scales = self.scales / self.scales.min()
        
        filter_rad = r_min
        n_neighbours = int(np.ceil(filter_rad))
        offset_range = cp.arange(-n_neighbours, n_neighbours + 1, dtype=cp.int32)
        
        a, b = cp.meshgrid(offset_range, offset_range, indexing='ij')
        offsets = cp.vstack([a.ravel(), b.ravel()]).T
        offsets_adjusted = offsets * self.scales[None]
        
        distances = cp.linalg.norm(offsets_adjusted, axis=1)
        weights = (r_min - distances) / r_min
        valid_mask = weights > 0
        offsets = offsets[valid_mask]
        weights = weights[valid_mask]
        weights /= weights.sum()
        
        self.weights = cp.array(weights, dtype=self.dtype)
        self.offsets = cp.array(offsets, dtype=cp.int32)
        
        self.normalizer = cp.zeros(self.shape[0], dtype=self.dtype)
        get_filter_weights_2D_cuda(self.nelx, self.nely, self.offsets, self.weights, self.normalizer)
    
    def _matvec(self, rho):
        v_out = cp.zeros_like(rho)
        apply_filter_2D_cuda(rho, v_out, self.nelx, self.nely, self.offsets, self.weights)
        return v_out
    
    def _rmatvec(self, rho):
        v_out = cp.zeros_like(rho)
        apply_filter_2D_transpose_cuda(rho, v_out, self.nelx, self.nely, self.offsets, self.weights, self.normalizer)
        return v_out
    
class CuGeneralFilter(FilterKernel):
    """
    CUDA-accelerated density filter for unstructured meshes.
    
    GPU version of GeneralFilter using CuPy sparse matrices. Implements density filtering
    for general unstructured meshes using spatial search. Stores explicit sparse filter matrix.
    
    Parameters
    ----------
    mesh : CuGeneralMesh
        CUDA-accelerated general unstructured mesh
    r_min : float
        Filter radius in physical units (not element units)
        
    Attributes
    ----------
    kernel : cupyx.scipy.sparse.csr_matrix
        Explicit filter matrix stored on GPU, shape (n_elements, n_elements)
    shape : tuple
        Filter matrix dimensions
        
    Methods
    -------
    dot(rho)
        Forward filter: H @ rho
    _rmatvec(sens)
        Adjoint filter: H^T @ sens
        
    Notes
    -----
    - Uses spatial search (KDTree on CPU) to build filter, then transfers to GPU
    - Stores explicit sparse matrix (memory-intensive for large meshes)
    - r_min is in physical units, not element-relative
    - Automatically detects 2D vs 3D from mesh
    - GPU acceleration provides 3-5x speedup over CPU for large meshes
    
    Examples
    --------
    >>> from pyFANTOM.CUDA import GeneralMesh, GeneralFilter
    >>> mesh = GeneralMesh(nodes, elements)  # Triangular mesh
    >>> filter = GeneralFilter(mesh=mesh, r_min=0.05)  # Physical radius
    >>> rho_filtered = filter.dot(rho)
    """
    def __init__(self, mesh: CuGeneralMesh, r_min):
        super().__init__()
        self.dtype = mesh.dtype
        self.nd = mesh.nodes.shape[1]
        if self.nd == 2:
            self.kernel = filter_kernel_2D_general(mesh.elements, mesh.centeroids, r_min)
        else:
            self.kernel = filter_kernel_3D_general(mesh.elements, mesh.centeroids, r_min)
        
        self.kernel = cp.sparse.csr_matrix(self.kernel, dtype=self.dtype)
        
        self.shape = self.kernel.shape 
        
        self.weights = cp.empty(self.shape[0], dtype=self.dtype)
        
    def _matvec(self, rho):
        """
        Forward filter application (internal).
        
        Parameters
        ----------
        rho : cp.ndarray
            Raw design variables on GPU, shape (n_elements,)
            
        Returns
        -------
        cp.ndarray
            Filtered design variables on GPU, shape (n_elements,)
        """
        return (self.kernel @ rho).reshape(rho.shape)
    
    def _rmatvec(self, rho):
        """
        Adjoint filter application (internal).
        
        Parameters
        ----------
        rho : cp.ndarray
            Filtered sensitivities on GPU, shape (n_elements,)
            
        Returns
        -------
        cp.ndarray
            Raw sensitivities on GPU, shape (n_elements,)
            
        Notes
        -----
        Used for backpropagating gradients through the filter in optimization.
        """
        return (self.kernel.T @ rho).reshape(rho.shape)